Minimize the NVD dumps
