+++ pkg/vulnloader/nvdloader/loader.go
@@ -14,6 +14,7 @@ import (
 	"github.com/stackrox/dotnet-scraper/types"
 	"github.com/stackrox/rox/pkg/httputil/proxy"
 	"github.com/stackrox/rox/pkg/utils"
+	"github.com/stackrox/scanner/pkg/nvdutils"
 	"github.com/stackrox/scanner/pkg/vulndump"
 	"github.com/stackrox/scanner/pkg/vulnloader"
 )
@@ -74,6 +75,9 @@ func downloadFeedForYear(enrichmentMap map[string]*types.FileFormat, outputDir s
 	}
 
 	for _, item := range dump.CVEItems {
+		if !nvdutils.CheckValidityAndTrim(item) {
+			continue
+		}
 		if enrichedEntry, ok := enrichmentMap[item.CVE.CVEDataMeta.ID]; ok {
 			// Add the CPE matches instead of removing for backwards compatibility purposes
 			item.Configurations.Nodes = append(item.Configurations.Nodes, &schema.NVDCVEFeedJSON10DefNode{
